# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FaisalOrakzai Digital OS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ActivityItem {
  id: number;
  type: string;
  title: string;
  description: string;
  createdAt: string;
}

export interface Project {
  id: number;
  title: string;
  category: string;
  description: string;
  status: string;
  year: number;
  /** @nullable */
  imageUrl?: string | null;
  technologies?: string[];
  /** @nullable */
  vision?: string | null;
  /** @nullable */
  roadmap?: string | null;
}

export interface Article {
  id: number;
  title: string;
  category: string;
  excerpt: string;
  content: string;
  readTime: number;
  publishedAt: string;
  /** @nullable */
  imageUrl?: string | null;
  tags?: string[];
}

export interface DashboardSummary {
  totalProjects: number;
  totalArticles: number;
  totalResearch: number;
  totalNews: number;
  totalCompanies: number;
  recentActivity: ActivityItem[];
  featuredProjects: Project[];
  featuredArticles: Article[];
}

export interface ProjectInput {
  title: string;
  category: string;
  description: string;
  status: string;
  year: number;
  imageUrl?: string;
  technologies?: string[];
  vision?: string;
  roadmap?: string;
}

export interface ArticleInput {
  title: string;
  category: string;
  excerpt: string;
  content: string;
  readTime: number;
  imageUrl?: string;
  tags?: string[];
}

export interface ResearchPaper {
  id: number;
  title: string;
  type: string;
  abstract: string;
  publishedAt: string;
  authors: string[];
  /** @nullable */
  downloadUrl?: string | null;
  tags?: string[];
}

export interface ResearchPaperInput {
  title: string;
  type: string;
  abstract: string;
  authors: string[];
  downloadUrl?: string;
  tags?: string[];
}

export interface NewsItem {
  id: number;
  title: string;
  category: string;
  excerpt: string;
  /** @nullable */
  content?: string | null;
  publishedAt: string;
  source: string;
  /** @nullable */
  imageUrl?: string | null;
}

export interface NewsItemInput {
  title: string;
  category: string;
  excerpt: string;
  content?: string;
  source: string;
  imageUrl?: string;
}

export interface Company {
  id: number;
  name: string;
  sector: string;
  description: string;
  founded: number;
  status: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  connections?: number[];
}

export interface MediaItem {
  id: number;
  title: string;
  type: string;
  description: string;
  publishedAt: string;
  /** @nullable */
  thumbnailUrl?: string | null;
  /** @nullable */
  videoUrl?: string | null;
  /** @nullable */
  duration?: string | null;
}

export interface TimelineEvent {
  id: number;
  year: number;
  title: string;
  description: string;
  type: string;
  isVision?: boolean;
}

export interface ContactInput {
  name: string;
  email: string;
  type: string;
  message: string;
  organization?: string;
}

export interface ContactSubmission {
  id: number;
  name: string;
  email: string;
  type: string;
  message: string;
  /** @nullable */
  organization?: string | null;
  createdAt: string;
}

export interface AiChatInput {
  message: string;
  /** @nullable */
  conversationId?: string | null;
}

export interface ChatReply {
  reply: string;
  conversationId: string;
}

export interface SearchResult {
  id: number;
  type: string;
  title: string;
  excerpt: string;
  url: string;
}

export interface SearchResults {
  query: string;
  results: SearchResult[];
}

export type ListProjectsParams = {
category?: string;
};

export type ListArticlesParams = {
category?: string;
search?: string;
};

export type ListResearchParams = {
type?: string;
};

export type ListNewsParams = {
category?: string;
};

export type ListMediaParams = {
type?: string;
};

export type GlobalSearchParams = {
q: string;
};