# lib/api-zod/src/generated/types/company.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FaisalOrakzai Digital OS API
 * OpenAPI spec version: 0.1.0
 */

export interface Company {
  id: number;
  name: string;
  sector: string;
  description: string;
  founded: number;
  status: string;
  /** @nullable */
  website?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  connections?: number[];
}