# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FaisalOrakzai Digital OS API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItem } from './activityItem';
import type { Article } from './article';
import type { Project } from './project';

export interface DashboardSummary {
  totalProjects: number;
  totalArticles: number;
  totalResearch: number;
  totalNews: number;
  totalCompanies: number;
  recentActivity: ActivityItem[];
  featuredProjects: Project[];
  featuredArticles: Article[];
}