# lib/api-zod/src/generated/types/researchPaperInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FaisalOrakzai Digital OS API
 * OpenAPI spec version: 0.1.0
 */

export interface ResearchPaperInput {
  title: string;
  type: string;
  abstract: string;
  authors: string[];
  downloadUrl?: string;
  tags?: string[];
}