# lib/api-zod/src/generated/types/timelineEvent.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * FaisalOrakzai Digital OS API
 * OpenAPI spec version: 0.1.0
 */

export interface TimelineEvent {
  id: number;
  year: number;
  title: string;
  description: string;
  type: string;
  isVision?: boolean;
}